Added variables, handlers, and conditionals
Have added conditionals, handlers, and variables to the 'playbook.yml'
file. Now, any packages that need to be installed have to be passed as
variables in the beginning of the YAML file. Also, depending on the type
of OS(Debian or Red Hat), the appropriate package manager will be used
for the installation of the packages.