New role for K8s installation on RPi
Have created a new role(s) for the preparation and installation of
Kubernetes on Raspberry Pi. The role 'base' prepares the Pis by
performing all actions needed for installation of K8s, and installs
required components of K8s, including the Weave networking driver.

The 'k8s-master' role initialises the Kubernetes master. This role
should be run against the node that should be the master node.

The 'k8s-minion' role initialises the Kubernetes minions and connects
them to the Kubernetes master using the join-token generated during the
initialisation of the master node.